# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HAIROU BTP Management Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /pointage:
    get:
      operationId: listPointageSheets
      tags: [pointage]
      summary: List pointage sheets
      security:
        - bearerAuth: []
      parameters:
        - name: projectId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of pointage sheets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PointageSheet"
    post:
      operationId: createPointageSheet
      tags: [pointage]
      summary: Create a pointage sheet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePointageSheetRequest"
      responses:
        "201":
          description: Pointage sheet created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PointageSheet"
  /pointage/{id}:
    get:
      operationId: getPointageSheet
      tags: [pointage]
      summary: Get pointage sheet by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Pointage sheet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PointageSheetDetail"
    put:
      operationId: updatePointageSheet
      tags: [pointage]
      summary: Update pointage sheet
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePointageSheetRequest"
      responses:
        "200":
          description: Pointage sheet updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PointageSheet"
  /pointage/{id}/submit:
    post:
      operationId: submitPointageSheet
      tags: [pointage]
      summary: Submit pointage sheet for approval
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitPointageRequest"
      responses:
        "200":
          description: Sheet submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PointageSheet"
  /pointage/{id}/approve:
    post:
      operationId: approvePointageSheet
      tags: [pointage]
      summary: Approve or reject pointage sheet
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApprovePointageRequest"
      responses:
        "200":
          description: Sheet approved/rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PointageSheet"
  /personnel:
    get:
      operationId: listPersonnel
      tags: [personnel]
      summary: List all personnel
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of personnel
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Personnel"
    post:
      operationId: createPersonnel
      tags: [personnel]
      summary: Create personnel record
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePersonnelRequest"
      responses:
        "201":
          description: Personnel created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Personnel"
  /personnel/{id}:
    put:
      operationId: updatePersonnel
      tags: [personnel]
      summary: Update personnel
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePersonnelRequest"
      responses:
        "200":
          description: Personnel updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Personnel"
    delete:
      operationId: deletePersonnel
      tags: [personnel]
      summary: Delete personnel
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Personnel deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    PointageEntry:
      type: object
      properties:
        id:
          type: integer
        personnelId:
          type: integer
        personnelName:
          type: string
        status:
          type: string
          enum: [PRESENT, ABSENT, DEMI_JOURNEE]
        arrivalTime:
          type: string
        departureTime:
          type: string
        hoursWorked:
          type: number
        dailyWage:
          type: number
        totalPay:
          type: number
        notes:
          type: string
      required:
        - id
        - personnelId
        - personnelName
        - status
    PointageSheet:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        projectName:
          type: string
        date:
          type: string
          format: date
        chefId:
          type: integer
        chefName:
          type: string
        status:
          type: string
          enum: [BROUILLON, SOUMISE, APPROUVEE, REJETEE, ARCHIVEE]
        signatureData:
          type: string
        adminComment:
          type: string
        totalPay:
          type: number
        createdAt:
          type: string
          format: date-time
        submittedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - date
        - status
    PointageSheetDetail:
      allOf:
        - $ref: "#/components/schemas/PointageSheet"
        - type: object
          properties:
            entries:
              type: array
              items:
                $ref: "#/components/schemas/PointageEntry"
    CreatePointageSheetRequest:
      type: object
      properties:
        projectId:
          type: integer
        date:
          type: string
          format: date
        entries:
          type: array
          items:
            $ref: "#/components/schemas/CreatePointageEntryRequest"
      required:
        - projectId
        - date
    CreatePointageEntryRequest:
      type: object
      properties:
        personnelId:
          type: integer
        status:
          type: string
          enum: [PRESENT, ABSENT, DEMI_JOURNEE]
        arrivalTime:
          type: string
        departureTime:
          type: string
        hoursWorked:
          type: number
        dailyWage:
          type: number
        notes:
          type: string
      required:
        - personnelId
        - status
    UpdatePointageSheetRequest:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/CreatePointageEntryRequest"
    SubmitPointageRequest:
      type: object
      properties:
        signatureData:
          type: string
      required:
        - signatureData
    ApprovePointageRequest:
      type: object
      properties:
        approved:
          type: boolean
        comment:
          type: string
      required:
        - approved
    Personnel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        trade:
          type: string
        phone:
          type: string
        idNumber:
          type: string
        emergencyContact:
          type: string
        dailyWage:
          type: number
        contractType:
          type: string
          enum: [CDI, CDD, JOURNALIER]
        projectIds:
          type: array
          items:
            type: integer
        totalDaysWorked:
          type: integer
        totalPayOwed:
          type: number
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - trade
        - dailyWage
        - contractType
        - isActive
    CreatePersonnelRequest:
      type: object
      properties:
        name:
          type: string
        trade:
          type: string
        phone:
          type: string
        idNumber:
          type: string
        emergencyContact:
          type: string
        dailyWage:
          type: number
        contractType:
          type: string
          enum: [CDI, CDD, JOURNALIER]
      required:
        - name
        - trade
        - dailyWage
        - contractType
    UpdatePersonnelRequest:
      type: object
      properties:
        name:
          type: string
        trade:
          type: string
        phone:
          type: string
        idNumber:
          type: string
        emergencyContact:
          type: string
        dailyWage:
          type: number
        contractType:
          type: string
          enum: [CDI, CDD, JOURNALIER]
        isActive:
          type: boolean
...