# lib/api-zod/src/generated/types/pointageEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HAIROU BTP Management Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { PointageEntryStatus } from "./pointageEntryStatus";

export interface PointageEntry {
  id: number;
  personnelId: number;
  personnelName: string;
  status: PointageEntryStatus;
  arrivalTime?: string;
  departureTime?: string;
  hoursWorked?: number;
  dailyWage?: number;
  totalPay?: number;
  notes?: string;
}